# Optimize sources processing perf on start