This PR:

- Speeds up the processing in `newSourcesToCompleteSourceDetails`:
  - Used object destructuring and manual lookup tables instead of Lodash
  - Changed `.forEach` to be `for...of` loops
  - Created lookup tables in a single loop instead of multiples
  - Used some `Object.assign` instead of spreads
- Speeds up `SourcesTree`'s `mapState`, by using an RTK `selectTotal` selector instead of calling `Object.values(sources).length` every time
- Consolidates the two `addSources` and `allSourcesReceived` actions that we were dispatching back-to-back, and removes storing the original source data in Redux (which we never read or process anywhere ever again anyway)

Based on local perf profiling, what I'm seeing is that the overall time to execute `setupDebugger` dropped from about 675ms to about 210ms.  More specifically, `newSourcesToCompleteSourceDetails` dropped from about 240ms to 105ms, and `SourcesTree.mapState` dropped from 250ms to, uh... small enough I can't see it in the profiler any more :)